added readme and restructured folders
