+++ README.md
@@ -1,13 +1,27 @@
 # Project Happy Thoughts API
 
-Replace this readme with your own information about your project.
-
-Start by briefly describing the assignment in a sentence or two. Keep it short and to the point.
+This is the backend API for a project called "Happy Thoughts," which allows users to post and like positive thoughts. The goal is to create a simple and uplifting platform for sharing joy.
 
 ## The problem
 
-Describe how you approached to problem, and what tools and techniques you used to solve it. How did you plan? What technologies did you use? If you had more time, what would be next?
+The project involves building a backend server using Node.js, Express, and MongoDB to handle the storage and retrieval of positive thoughts. The provided code includes routes for getting all thoughts, posting new thoughts, and liking existing thoughts.
+
+## Approach
+Project Structure: The code is organized into different files for better maintainability. There's a models folder for defining the Mongoose model, a routes folder for handling different API endpoints, and the main server.js file to tie everything together.
+
+Database: MongoDB is used as the database, and Mongoose is employed as an ODM (Object Data Modeling) library to interact with the database.
+
+Route Handling: Express.js is utilized to create routes that handle various HTTP requests. For example, there are routes to get all thoughts, post a new thought, and like a thought.
+
+Error Handling: There's error handling in place using try-catch blocks to ensure that if something goes wrong, the server responds with the appropriate status code and error message.
+
+## Technologies Used
+
+Node.js
+Express.js
+MongoDB (with Mongoose)
+CORS (Cross-Origin Resource Sharing)
 
 ## View it live
 
-Every project should be deployed somewhere. Be sure to include the link to the deployed project so that the viewer can click around and see what it's all about.
+https://api-happy-thought.onrender.com/
